Fix @GetMapping @PostMapping @DeleteMapping Replace getById() with findById().get() Disable CSRF Set antMatchers for Http method

+++ src/main/java/com/thanos/SecurityDemo/Repository/Repo.java
@@ -2,7 +2,8 @@ package com.thanos.SecurityDemo.Repository;
 
 import com.thanos.SecurityDemo.Entity.Customer;
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
 
 
-public interface Repo extends JpaRepository<Customer,Long> {
+public interface Repo extends JpaRepository<Customer, Integer> {
 }
